Specify path which we'll load bundler from
Fixes #456. See my analysis at
https://github.com/rails/spring/issues/456#issuecomment-177282363

Previously we were implicitly relying on Bundler setting RUBYLIB so that
we'd be able to find Bundler in the child process.

Bundler 1.11 no longer does that for us, so we need to explicitly do it
ourselves (it was a bug that Bundle.with_clean_env didn't clear it).