Run each command in its own process group
Before this change, a command would run in the same process group as the
long-running application process.

When a user runs a command and then hits e.g. ^C this causes a SIGINT.
The code in Spring::Client::Run#forward_signals causes the SIGINT to be
trapped and then forwarded to the pid which is running the actual
command.

However, we don't forward to the pid but to the pgid (process group id).
This is so that if the process we're running creates child processes in
the same process group, those children should also receive the signal.

So before this change, the long-running application process *and* the
short-running process which runs the individual command would both
receive a SIGINT, because they shared a pgid.

However the long-running process would ignore the signal due to the code
in Spring::Server#ignore_signals. Therefore it wouldn't do anything with
the SIGINT (which is good, because we don't want the long-running
process to exit).

Now, when the server is run via "spring server", we don't run
Spring::Server#ignore_signals, because the server runs in the foreground
and we *do* want it to exit if the user types ^C on the command line.
But this means that the long-running application process *would* receive
a SIGINT in the scenario above, which is bad.

The solution is to ensure that when we fork for an individual command
run, we create a new process group.