Forward signals to the application over the socket
Rather than assuming that the client process can directly signal the
application process (i.e. they are within the same pid namespace).

This should make it easier to communicate with a spring server which is
inside a Docker container (especially if the Docker host is not the same
as the host running the spring client).