Refer to Spring as a proper noun, with a capital "S" (#559)
This updates the docs and code - particularly the comments
and code that generate help, e.g. when you run `spring --help` -
to refer to "Spring" with a capital "S", since it's a name (i.e.
a proper noun).

The only places where I've left "spring" with a lower-case "s"
is where we're referring to the name of the gem (e.g. in a
`Gemfile`) or for code examples (e.g. running `bin/spring`).